[dlocate] separate the program into different modules
