+++ dlocate/README.md
@@ -1,0 +1,6 @@
+# dlocate utility
+
+run using this command
+```
+go run searh.go data.go main.go /home/ahmed/Downloads/csed/networks
+```
